requirements.txt
# ===== Core =====
numpy==1.26.2
pandas==2.3.3
pydantic==2.11.10

# ===== NLP Basics =====
nltk==3.9.2
regex==2025.11.3
tqdm==4.67.1
textblob==0.15.3

# ===== Arabic NLP =====
PyArabic==0.6.15
arabic-reshaper==3.0.0
python-bidi==0.6.7

# ===== Jupyter / Dev =====
jupyter==1.1.1
matplotlib==3.10.8
seaborn==0.13.2
wordcloud==1.9.5


# ===== Fast api ======
fastapi==0.128.0
uvicorn==0.40.0
# torch==2.2.2+cpu
transformers==4.57.3

# ===== Celery =====
celery==5.6.2
redis==7.1.0
openpyxl==3.1.5
python-multipart==0.0.21
arabic-reshaper==3.0.0
python-bidi==0.6.7
#=========env=========
python-dotenv==1.2.1
pydantic-settings==2.12.0
pydantic==2.12.5











